23-üK294
To keep things simple and clean. A entire new project was started and most of the code was copied from the files before to speed up the process and reuse already working code. This new project was created because other files got corrupt and the code that I've used is mostly out of Git History. By interests about this process please check the history of different branches and not only main.